backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
scikit-image
numpy
fpdf
Pillow
matplotlib 